refactor(cache): split cache.go into keygen/vary/cachecontrol/utils files
cache.go had grown past 1300 lines. Extract cohesive groups into separate
files within the same package (no behavior change, no exported-API change):

- keygen.go: defaultKeyGenerator, the canonical query/header/cookie helpers,
  escapeKeyDelimiters, the segment bounding/hashing helpers, and the
  key-tuning constants plus the key buffer pool.
- vary.go: Vary parsing, the vary-key hasher, manifest load/store, and the
  maxVaryHeaders cap (now next to its only user).
- cachecontrol.go: Cache-Control directive parsing and interpretation.
- utils.go: age/freshness, HTTP-date and seconds helpers, and the auth hasher.

defaultKeyGenerator is also simplified: the QUERY request-body branch moves
into appendQueryBodySegment, and the "return buffer to pool" boilerplate
becomes the (inlined) releaseKeyBuffer helper. The pool "get" is kept inline
because a helper for it does not inline and sits in the hot path.

Output is byte-identical (Test_defaultKeyGenerator_stableKeys) and the key
benchmark shows unchanged allocations, so cache keys are unaffected.

cache.go: ~1372 -> 891 lines.
